requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-dotenv>=1.0.0
httpx>=0.28.0
pytest>=8.3.0
google-cloud-storage>=2.10.0
jinja2>=3.1.0
python-multipart>=0.0.6
pytest-asyncio>=0.21.0

# Transformer Service Dependencies
pdfplumber>=0.10.3
pypandoc>=1.12
python-docx>=1.1.0
python-pptx>=0.6.23
openpyxl>=3.1.2
xlrd>=2.0.1
odfpy>=1.4.1
striprtf>=0.0.26
google-cloud-speech>=2.21.0
structlog>=23.2.0
tenacity>=8.2.0
PyYAML>=6.0.1

# Tabular Service Dependencies
pandas>=2.0.0
pyarrow>=12.0.0
pandera>=0.17.0
python-Levenshtein>=0.21.0  # For fuzzy string matching
rapidfuzz>=3.0.0  # Faster alternative to python-Levenshtein
google-cloud-bigquery>=3.11.0
openai>=1.0.0  # For Featherless.ai API (LLM)

# CPU-only PyTorch (MUST be installed BEFORE sentence-transformers to avoid NVIDIA deps)
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.0.0+cpu
sentence-transformers>=2.3.0  # For local embeddings (paraphrase-multilingual-mpnet-base-v2)

numpy>=1.24.0
scikit-learn>=1.3.0  # For cosine_similarity
